+++ server/snapshot/createSnapshot.ts
@@ -6,16 +6,11 @@ import isBinaryPath from 'is-binary-path'
 import { Answers, saoPrompts } from '../common/prompts'
 import { generate } from '../service/generate'
 
-const excludedList = {
-  prismaDB: ['postgresql'],
-  typeormDB: ['postgres', 'mongodb', 'mssql', 'mariadb', 'cockroachdb']
-}
-
 const createAnswersList = (
   answers: Answers,
   [nextQuestion, ...leftQuestions]: typeof saoPrompts
 ): Answers[] => {
-  if (nextQuestion.name === 'pm' || nextQuestion.when?.(answers) === false) {
+  if (nextQuestion.when?.(answers) === false) {
     return leftQuestions.length
       ? createAnswersList(answers, leftQuestions)
       : [answers]
@@ -36,26 +31,17 @@ const createAnswersList = (
     return leftQuestions.length ? createAnswersList(ans, leftQuestions) : [ans]
   }
 
-  return nextQuestion.choices
-    .filter(
-      (choice) =>
-        !excludedList[nextQuestion.name as keyof typeof excludedList]?.includes(
-          choice.value
-        )
-    )
-    .reduce<Answers[]>((prev, choice) => {
-      const ans = {
-        ...answers,
-        dir: `${answers.dir}${answers.dir ? '-' : ''}${choice.value}`,
-        [nextQuestion.name]: choice.value
-      }
-      return [
-        ...prev,
-        ...(leftQuestions.length
-          ? createAnswersList(ans, leftQuestions)
-          : [ans])
-      ]
-    }, [])
+  return nextQuestion.choices.reduce<Answers[]>((prev, choice) => {
+    const ans = {
+      ...answers,
+      dir: `${answers.dir}${answers.dir ? '-' : ''}${choice.value}`,
+      [nextQuestion.name]: choice.value
+    }
+    return [
+      ...prev,
+      ...(leftQuestions.length ? createAnswersList(ans, leftQuestions) : [ans])
+    ]
+  }, [])
 }
 
 const listFiles = (targetDir: string): string[] =>
@@ -67,8 +53,16 @@ const listFiles = (targetDir: string): string[] =>
     }, [])
 
 export const createSnapshot = async (rootDir: string) => {
-  const promptsList = createAnswersList({ dir: '' }, saoPrompts.slice(1))
   const outputDir = path.join(os.tmpdir(), `${Date.now()}`)
+  const appPromptsList = createAnswersList({ dir: '' }, saoPrompts.slice(1, -2))
+  const promptsList = [
+    ...appPromptsList,
+    ...createAnswersList(appPromptsList[0], saoPrompts.slice(-2)),
+    ...createAnswersList(
+      appPromptsList[appPromptsList.length - 1],
+      saoPrompts.slice(-2)
+    )
+  ]
 
   console.log(promptsList.length)
   fs.mkdirSync(outputDir)
@@ -87,14 +81,38 @@ export const createSnapshot = async (rootDir: string) => {
     )
   }
 
-  const allFiles = listFiles(outputDir).sort()
-  console.log(allFiles.length)
-  let text = ''
-  for (let i = 0; i < allFiles.length; i += 1) {
-    text += `=== ${allFiles[i].replace(`${outputDir}/`, '')} ===
-${isBinaryPath(allFiles[i]) ? 'binary' : fs.readFileSync(allFiles[i], 'utf8')}
-`
-  }
+  const hasDataList: { name: string; data: string }[] = []
+  const fileData = listFiles(outputDir)
+    .sort()
+    .map((file) => {
+      const data = isBinaryPath(file)
+        ? '\nbinary\n'
+        : `\n\n\`\`\`\n${fs
+            .readFileSync(file, 'utf8')
+            .replace(
+              /```/g,
+              path.extname(file) === '.md' ? '\\```' : '```'
+            )}\n\`\`\`\n`
+      const name = file.replace(`${outputDir}/`, '')
+      const duplicated = hasDataList.find(
+        (d) =>
+          d.name.split('/').pop() === file.split('/').pop() && d.data === data
+      )
+
+      if (duplicated) {
+        return {
+          name: `[${name}](#${duplicated.name})`,
+          data: ''
+        }
+      } else {
+        const d = { name, data }
+        hasDataList.push(d)
+        return d
+      }
+    })
+
   rimraf.sync(outputDir)
-  return text
+
+  console.log(hasDataList.length, fileData.length)
+  return fileData.reduce((prev, d) => `${prev}### ${d.name}${d.data}\n`, '')
 }