+++ server/templates/client/nuxt/plugins/api.ts
@@ -25,6 +25,7 @@ declare module '@nuxt/types' {
 }
 
 declare module 'vuex/types/index' {
+  // eslint-disable-next-line no-unused-vars, @typescript-eslint/no-unused-vars
   interface Store<S> {
     $api: ApiInstance
   }