+++ server/templates/core/fastify-typeorm/server/index.ts
@@ -4,11 +4,12 @@ import Fastify from 'fastify'
 import helmet from 'fastify-helmet'
 import cors from 'fastify-cors'
 import fastifyStatic from 'fastify-static'
-import fastifyAuth from 'fastify-auth'
+import fastifyJwt from 'fastify-jwt'
 import { createConnection } from 'typeorm'
 import server from './$server'
 import ormOptions from './$orm'
 import {
+  JWT_SECRET,
   SERVER_PORT,
   BASE_PATH,
   TYPEORM_HOST,
@@ -26,9 +27,9 @@ fastify.register(fastifyStatic, {
   root: path.join(__dirname, 'public'),
   prefix: BASE_PATH
 })
-fastify.register(fastifyAuth).after(() => {
-  server(fastify, { basePath: BASE_PATH })
-})
+fastify.register(fastifyJwt, { secret: JWT_SECRET })
+
+server(fastify, { basePath: BASE_PATH })
 
 createConnection({
   type: '<%= typeormDB %>',