+++ server/templates/testing/jest/server/test/server.test.ts
@@ -1,3 +1,4 @@
+import <%= server %> from '<%= server %>'
 import controller from '$/api/tasks/controller'
 import { getTasks } from '$/service/tasks'
 
@@ -26,21 +27,22 @@ test('dependency injection into controller', async () => {
             { id: 3, label: 'task4', done: true },
             { id: 4, label: 'task5', done: false }
           ])
-      })<% } else { %>readDB: () => Promise.resolve({
-        nextId: 5,
-        tasks: [
-          { id: 0, label: 'task1', done: false },
-          { id: 1, label: 'task2', done: false },
-          { id: 2, label: 'task3', done: true },
-          { id: 3, label: 'task4', done: true },
-          { id: 4, label: 'task5', done: false }
-        ]
-      })<% } %>
+      })<% } else { %>readDB: () =>
+        Promise.resolve({
+          nextId: 5,
+          tasks: [
+            { id: 0, label: 'task1', done: false },
+            { id: 1, label: 'task2', done: false },
+            { id: 2, label: 'task3', done: true },
+            { id: 3, label: 'task4', done: true },
+            { id: 4, label: 'task5', done: false }
+          ]
+        })<% } %>
     }),
     print: (text: string) => {
       printedMessage = text
     }
-  })()
+  })(<%= server %>())
 
   const limit = 3
   const message = 'test message'